+++ packages/badge/src/Badge/Badge.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import { cx } from '@leafygreen-ui/emotion';
 import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
 
-import { badgeVariants, baseStyle } from './styles';
-import { BadgeProps, Variant } from './types';
+import { badgeVariants, baseStyle } from './Badge.styles';
+import { BadgeProps, Variant } from './Badge.types';
 
 /**
  * Badges can be used to highlight information or the status of something.