+++ packages/marketing-modal/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @leafygreen-ui/marketing-modal
 
+## 6.0.0
+
+### Major Changes
+
+- 0757cfbfc: Updates Typescript build to TS5.8
+
+### Patch Changes
+
+- 0757cfbfc: Adds `@lg-tools/build` as a dev dependency
+- 0757cfbfc: Adds missing `@lg-tools/` devDependencies.
+  Updates `build`, `tsc` & `docs` scripts to use `lg-build *` cli
+- 0757cfbfc: Updates `types` entry point to `./dist/types`.
+  Removes redundant `compilerOptions` from TSConfig
+- 0757cfbfc: Updates `main` entry point in package.json to `./dist/umd`
+- Updated dependencies [0757cfbfc]
+- Updated dependencies [0757cfbfc]
+- Updated dependencies [0757cfbfc]
+- Updated dependencies [0757cfbfc]
+- Updated dependencies [0757cfbfc]
+- Updated dependencies [0757cfbfc]
+  - @leafygreen-ui/leafygreen-provider@5.0.0
+  - @leafygreen-ui/typography@21.0.0
+  - @leafygreen-ui/emotion@5.0.0
+  - @leafygreen-ui/palette@5.0.0
+  - @leafygreen-ui/button@24.0.0
+  - @leafygreen-ui/tokens@3.0.0
+  - @leafygreen-ui/modal@18.0.0
+  - @leafygreen-ui/lib@15.0.0
+
+## 5.0.15
+
+### Patch Changes
+
+- @leafygreen-ui/button@23.1.6
+- @leafygreen-ui/modal@17.1.7
+- @leafygreen-ui/typography@20.1.9
+
 ## 5.0.14
 
 ### Patch Changes